requirements.txt
requests
pyyaml
pprint
ratelimit